Adds xls processing methods from group controller into a helper file
